Revert "Revert "Switched eval_t from float to int16_t""
This reverts commit 0a940c23e3c5ca30331794c771aadbefc871c14d.

+++ Typedefs.h
@@ -5,4 +5,4 @@ using move_t = uint16_t;
 using zobrist_t = uint64_t;
 using see_mask = int32_t;
 using packed_eval_t = uint64_t;
-using eval_t = float;
+using eval_t = int16_t;
